requirements.txt
cassandra-driver==3.18.0
PyYAML==5.1.1
